+++ app/src/main/java/id/rllyhz/githubuserapp/data/response/SearchUsersResponse.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
 data class SearchUsersResponse(
-    @SerializedName("total_count") @Expose val totalResults: Int,
+    @SerializedName("total_count") @Expose val totalResults: Long,
     @SerializedName("items") @Expose val users: List<UsersResponse>
 )
